rag-chatbot/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-qdrant
langchain-huggingface
qdrant-client
sentence-transformers
rank-bm25
fastapi
uvicorn
python-dotenv
httpx
